refactor: split org context and add logger
