+++ src/lib/logger.js
@@ -1,0 +1,7 @@
+const logger = {
+  log: (...args) => console.log(...args),
+  error: (...args) => console.error(...args),
+  warn: (...args) => console.warn(...args)
+};
+
+export default logger;
